Revert "Merge pull request #2637 from ousttrue/fix/node_name_unique"
This reverts commit ed5988e25a88476fe6da61a0000ae35dec7a3288, reversing
changes made to cc1f8c0fc94cd7ed5b1354ef4f01e5183e959b4b.

# Conflicts:
#	Assets/UniGLTF/Runtime/UniGLTF/IO/GltfData.cs
#	Assets/UniGLTF/Runtime/UniGLTF/IO/GltfDuplicatedNameConversionRule.cs
#	Assets/UniGLTF/Runtime/UniGLTF/IO/ImporterContext.cs
