+++ Assets/UniGLTF/Tests/UniGLTF/GlbParserTests.cs
@@ -35,8 +35,6 @@ namespace UniGLTF
             var parser = new GlbLowLevelParser("Test", data.ToGlbBytes());
             using (var parsed = parser.Parse())
             {
-                GltfDuplicatedNameConversionRule.FixNames(parsed.GLTF);
-
                 Assert.AreEqual("FooBar", parsed.GLTF.textures[0].name);
                 // NOTE: 大文字小文字が違うだけの名前は、同一としてみなされ、Suffix が付く。
                 Assert.AreEqual("foobar__UNIGLTF__DUPLICATED__2", parsed.GLTF.textures[1].name);
