+++ Assets/VRM10/Runtime/IO/Vrm10Importer.cs
@@ -79,15 +79,6 @@ namespace UniVRM10
             m_springboneRuntime = springboneRuntime;
         }
 
-        protected override async Task PreprocessAsync(IAwaitCaller awaitCaller)
-        {
-            await base.PreprocessAsync(awaitCaller);
-            await awaitCaller.Run(() =>
-            {
-                GltfDuplicatedNameConversionRule.FixNodeNameUnique(GLTF);
-            });
-        }
-
         static void AssignHumanoid(List<VrmLib.Node> nodes, UniGLTF.Extensions.VRMC_vrm.HumanBone humanBone, VrmLib.HumanoidBones key)
         {
             if (nodes == null)
